// front/src/types/SiteStudyBasicGenericSectionFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: SiteStudyBasicGenericSectionFragment
// ====================================================

export interface SiteStudyBasicGenericSectionFragment {
  __typename: "SiteStudyBasicGenericSection";
  hide: boolean;
  title: string;
  name: string;
}